requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
decorator==4.4.2
Flask==1.1.2
future==0.18.2
geocoder==1.38.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
pycparser==2.20
pycscope==0.3
ratelim==0.1.6
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
